slice_provider/service-orchestrator-adaptor/requirements.txt
flask
flask-restful
requests
pyyaml
